Add 17 pigments to map
Thank you to Sumokau, Xared, Diarmuid for providing photos of the locations of pigments :D

I went in and found all the coordinates based on the images people took, which is why I don't have all the coordinates as I couldn't find everywhere :sob:

And with the sandcavern, I haven't generated renders for every zone whoops. I didn't realize there were pigments in other parts of the cave.